FIX: Organize Protected Routes and Fix Routing Issues for Unified Portal (#836)
* Display UnstractAndLlmWhispererLogo in login page, if available

* Refactor import of WhispererLogo to point to new path

* Refactored routes and fixed related bugs

* Rename logo name
