+++ frontend/src/components/helpers/auth/PersistentLogin.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { Outlet } from "react-router-dom";
+import { Outlet, useLocation } from "react-router-dom";
 
 import useSessionValid from "../../../hooks/useSessionValid";
 import { useSessionStore } from "../../../store/session-store";
@@ -9,8 +9,15 @@ import { PromptRun } from "../../custom-tools/prompt-card/PromptRun";
 
 let selectedProductStore;
 let selectedProduct;
+let setSelectedProduct;
+let SELECTED_PRODUCT;
+let PRODUCT_NAMES = {};
 try {
   selectedProductStore = require("../../../plugins/llm-whisperer/store/select-product-store.js");
+  SELECTED_PRODUCT =
+    require("../../../plugins/llm-whisperer/helper").SELECTED_PRODUCT;
+  PRODUCT_NAMES =
+    require("../../../plugins/llm-whisperer/helper").PRODUCT_NAMES;
 } catch {
   // Ignore if hook not available
 }
@@ -19,16 +26,23 @@ function PersistentLogin() {
   const [isLoading, setIsLoading] = useState(true);
   const { sessionDetails } = useSessionStore();
   const checkSessionValidity = useSessionValid();
+  const location = useLocation();
+  const queryParams = new URLSearchParams(location.search);
+  const selectedProductQueryParam = queryParams.get(SELECTED_PRODUCT);
 
   try {
     if (selectedProductStore?.useSelectedProductStore) {
       selectedProduct = selectedProductStore?.useSelectedProductStore(
         (state) => state?.selectedProduct
       );
+      setSelectedProduct = selectedProductStore.useSelectedProductStore(
+        (state) => state?.setSelectedProduct
+      );
     }
   } catch (error) {
     // Do nothing
   }
+
   useEffect(() => {
     let isMounted = true;
 
@@ -49,6 +63,15 @@ function PersistentLogin() {
 
     return () => (isMounted = false);
   }, [selectedProduct]);
+
+  useEffect(() => {
+    if (
+      selectedProductQueryParam &&
+      Object.values(PRODUCT_NAMES).includes(selectedProductQueryParam)
+    ) {
+      setSelectedProduct(selectedProductQueryParam);
+    }
+  }, [selectedProductQueryParam]);
 
   if (isLoading) {
     return <GenericLoader />;
