+++ frontend/src/routes/useMainAppRoutes.js
@@ -20,7 +20,6 @@ import { CustomToolsHelper } from "../components/helpers/custom-tools/CustomTool
 import { ToolIdePage } from "../pages/ToolIdePage.jsx";
 import { OutputAnalyzerPage } from "../pages/OutputAnalyzerPage.jsx";
 import { deploymentTypes } from "../helpers/GetStaticData.js";
-import { RequireAuth } from "../components/helpers/auth/RequireAuth.js";
 
 let RequirePlatformAdmin;
 let PlatformAdminPage;
@@ -30,6 +29,9 @@ let ChatAppLayout;
 let ManualReviewSettings;
 let OnboardProduct;
 let PRODUCT_NAMES = {};
+let ManualReviewPage;
+let SimpleManualReviewPage;
+let ReviewLayout;
 
 try {
   RequirePlatformAdmin =
@@ -66,9 +68,20 @@ try {
   // Do nothing.
 }
 
+try {
+  ManualReviewPage =
+    require("../plugins/manual-review/page/ManualReviewPage.jsx").ManualReviewPage;
+  ReviewLayout =
+    require("../plugins/manual-review/review-layout/ReviewLayout.jsx").ReviewLayout;
+  SimpleManualReviewPage =
+    require("../plugins/manual-review/page/simple/SimpleManualReviewPage.jsx").SimpleManualReviewPage;
+} catch (err) {
+  // Do nothing, Not-found Page will be triggered.
+}
+
 function useMainAppRoutes() {
   const routes = (
-    <Route path="" element={<RequireAuth />}>
+    <>
       <Route path=":orgName" element={<FullPageLayout />}>
         <Route path="onboard" element={<OnBoardPage />} />
       </Route>
@@ -140,7 +153,31 @@ function useMainAppRoutes() {
           </Route>
         )}
       </Route>
-    </Route>
+      {ReviewLayout && ManualReviewPage && (
+        <Route path=":orgName" element={<ReviewLayout />}>
+          <Route
+            path="review"
+            element={<ManualReviewPage type="review" />}
+          ></Route>
+          <Route
+            path="simple_review/review"
+            element={<SimpleManualReviewPage type="simple_review" />}
+          ></Route>
+          <Route
+            path="simple_review/approve"
+            element={<SimpleManualReviewPage type="simple_approve" />}
+          ></Route>
+          <Route
+            path="review/download_and_sync"
+            element={<ManualReviewPage type="download" />}
+          />
+          <Route
+            path="review/approve"
+            element={<ManualReviewPage type="approve" />}
+          />
+        </Route>
+      )}
+    </>
   );
 
   if (OnboardProduct && Object.keys(PRODUCT_NAMES)?.length) {
